first refactor steps
