+++ src/components/Proposal/index.ts
@@ -1,0 +1,6 @@
+export { default as History } from './History';
+export { default as Calls } from './Calls';
+export { default as Votes } from './Votes';
+export { default as Stakes } from './Stakes';
+export { default as Details } from './Details';
+export { default as Status } from './Status';
